+++ lib/php/lib/Type/TType.php
@@ -27,23 +27,23 @@ namespace Thrift\Type;
  */
 final class TType
 {
-    public const int STOP   = 0;
-    public const int VOID   = 1;
-    public const int BOOL   = 2;
-    public const int BYTE   = 3;
-    public const int I08    = 3;
-    public const int DOUBLE = 4;
-    public const int I16    = 6;
-    public const int I32    = 8;
-    public const int I64    = 10;
-    public const int STRING = 11;
-    public const int UTF7   = 11;
-    public const int STRUCT = 12;
-    public const int MAP    = 13;
-    public const int SET    = 14;
-    public const int LST    = 15;    // N.B. cannot use LIST keyword in PHP!
-    public const int UTF8   = 16;
-    public const int UTF16  = 17;
+    public const STOP   = 0;
+    public const VOID   = 1;
+    public const BOOL   = 2;
+    public const BYTE   = 3;
+    public const I08    = 3;
+    public const DOUBLE = 4;
+    public const I16    = 6;
+    public const I32    = 8;
+    public const I64    = 10;
+    public const STRING = 11;
+    public const UTF7   = 11;
+    public const STRUCT = 12;
+    public const MAP    = 13;
+    public const SET    = 14;
+    public const LST    = 15;    // N.B. cannot use LIST keyword in PHP!
+    public const UTF8   = 16;
+    public const UTF16  = 17;
 
     private function __construct()
     {